+++ fe/src/page/member/Member.css
@@ -31,7 +31,7 @@
 /*----------------MyPage 상위 여백-------------------------*/
 
 .member {
-    margin-top: 40px;
+    margin-top: 35px;
 }
 
 /*------------------- 화면 크기가 768px 이하 (태블릿 화면) -------------------*/
